pyproject.toml
[tool.poetry]
name = "costa-intro"
version = "0.1.0"
description = ""
authors = ["Luca Mattiazzi <l.d.mattiazzi@gmail.com>"]
readme = "README.md"
packages = [{ include = "costa_intro" }]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.1"
numpy = "^1.24.1"
openai-whisper = "^20230117"
unidecode = "^1.3.6"
nltk = "^3.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
